Cleanup: move unit-related functions to own translation unit
Create a units.c translation unit.
